+++ x-pack/plugins/fleet/server/services/agent_policy.test.ts
@@ -9,17 +9,13 @@ import { SavedObjectsErrorHelpers } from '@kbn/core/server';
 import { securityMock } from '@kbn/security-plugin/server/mocks';
 import { loggerMock } from '@kbn/logging-mocks';
 import type { Logger } from '@kbn/core/server';
-
 import type { SavedObjectError } from '@kbn/core-saved-objects-common';
-
-import axios from 'axios';
 
 import {
   PackagePolicyRestrictionRelatedError,
   FleetUnauthorizedError,
   HostedAgentPolicyRestrictionRelatedError,
   AgentPolicyInvalidError,
-  AgentlessAgentCreateError,
 } from '../errors';
 import type {
   AgentPolicy,
@@ -44,8 +40,6 @@ import * as outputsHelpers from './agent_policies/outputs_helpers';
 import { auditLoggingService } from './audit_logging';
 import { licenseService } from './license';
 import type { UninstallTokenServiceInterface } from './security/uninstall_token_service';
-import { listEnrollmentApiKeys } from './api_keys';
-import { listFleetServerHosts } from './fleet_server_host';
 
 function getSavedObjectMock(agentPolicyAttributes: any) {
   const mock = savedObjectsClientMock.create();
@@ -108,12 +102,6 @@ const mockedPackagePolicyService = packagePolicyService as jest.Mocked<typeof pa
 const mockedGetFullAgentPolicy = getFullAgentPolicy as jest.Mock<
   ReturnType<typeof getFullAgentPolicy>
 >;
-const mockedListEnrollmentApiKeys = listEnrollmentApiKeys as jest.Mock<
-  ReturnType<typeof listEnrollmentApiKeys>
->;
-const mockedListFleetServerHosts = listFleetServerHosts as jest.Mock<
-  ReturnType<typeof listFleetServerHosts>
->;
 
 function getAgentPolicyCreateMock() {
   const soClient = savedObjectsClientMock.create();
@@ -147,216 +135,6 @@ describe('Agent policy', () => {
 
   afterEach(() => {
     jest.resetAllMocks();
-  });
-
-  describe('createAgentlessAgent', () => {
-    beforeEach(() => {
-      (axios as jest.MockedFunction<typeof axios>).mockReset();
-      jest.clearAllMocks();
-    });
-
-    it('should throw AgentlessAgentCreateError if agentless policy does not support_agentless', async () => {
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: true } as any);
-
-      await expect(
-        agentPolicyService.createAgentlessAgent(esClient, soClient, {
-          id: 'mocked',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: false,
-        } as AgentPolicy)
-      ).rejects.toThrowError(
-        new AgentlessAgentCreateError('Agentless agent policy does not have agentless enabled')
-      );
-    });
-
-    it('should throw AgentlessAgentCreateError if cloud is not enabled', async () => {
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: false } as any);
-
-      await expect(
-        agentPolicyService.createAgentlessAgent(esClient, soClient, {
-          id: 'mocked',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: true,
-        } as AgentPolicy)
-      ).rejects.toThrowError(new AgentlessAgentCreateError('Agentless agent not supported'));
-    });
-
-    it('should throw AgentlessAgentCreateError if agentless configuration is not found', async () => {
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getConfig').mockReturnValue({} as any);
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: true } as any);
-
-      await expect(
-        agentPolicyService.createAgentlessAgent(esClient, soClient, {
-          id: 'mocked',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: true,
-        } as AgentPolicy)
-      ).rejects.toThrowError(new AgentlessAgentCreateError('missing agentless configuration'));
-    });
-    it('should throw AgentlessAgentCreateError if fleet hosts are not found', async () => {
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getConfig').mockReturnValue({
-        agentless: {
-          api: {
-            url: 'http://api.agentless.com/api/v1/ess',
-            tls: {
-              certificate: '/path/to/cert',
-              key: '/path/to/key',
-            },
-          },
-        },
-      } as any);
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: true } as any);
-      mockedListFleetServerHosts.mockResolvedValue({ items: [] } as any);
-      mockedListEnrollmentApiKeys.mockResolvedValue({
-        items: [
-          {
-            id: 'mocked',
-            policy_id: 'mocked',
-            api_key: 'mocked',
-          },
-        ],
-      } as any);
-
-      await expect(
-        agentPolicyService.createAgentlessAgent(esClient, soClient, {
-          id: 'mocked',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: true,
-        } as AgentPolicy)
-      ).rejects.toThrowError(new AgentlessAgentCreateError('missing Fleet server host'));
-    });
-
-    it('should throw AgentlessAgentCreateError if enrollment tokens are not found', async () => {
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getConfig').mockReturnValue({
-        agentless: {
-          api: {
-            url: 'http://api.agentless.com/api/v1/ess',
-            tls: {
-              certificate: '/path/to/cert',
-              key: '/path/to/key',
-            },
-          },
-        },
-      } as any);
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: true } as any);
-      mockedListFleetServerHosts.mockResolvedValue({
-        items: [
-          {
-            id: 'mocked',
-            host: 'http://fleetserver:8220',
-            active: true,
-            is_default: true,
-          },
-        ],
-      } as any);
-      mockedListEnrollmentApiKeys.mockResolvedValue({
-        items: [],
-      } as any);
-
-      await expect(
-        agentPolicyService.createAgentlessAgent(esClient, soClient, {
-          id: 'mocked',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: true,
-        } as AgentPolicy)
-      ).rejects.toThrowError(new AgentlessAgentCreateError('missing Fleet enrollment token'));
-    });
-
-    it('should create agentless agent', async () => {
-      const returnValue = {
-        id: 'mocked',
-        regional_id: 'mocked',
-      };
-
-      (axios as jest.MockedFunction<typeof axios>).mockResolvedValueOnce(returnValue);
-      const soClient = getAgentPolicyCreateMock();
-      // ignore unrelated unique name constraint
-      const esClient = elasticsearchServiceMock.createClusterClient().asInternalUser;
-      jest.spyOn(appContextService, 'getConfig').mockReturnValue({
-        agentless: {
-          api: {
-            url: 'http://api.agentless.com/api/v1/ess',
-            tls: {
-              certificate: '/path/to/cert',
-              key: '/path/to/key',
-              ca: '/path/to/ca',
-            },
-          },
-        },
-      } as any);
-      jest.spyOn(appContextService, 'getCloud').mockReturnValue({ isCloudEnabled: true } as any);
-      jest
-        .spyOn(appContextService, 'getKibanaVersion')
-        .mockReturnValue('mocked-kibana-version-infinite');
-      mockedListFleetServerHosts.mockResolvedValue({
-        items: [
-          {
-            id: 'mocked-fleet-server-id',
-            host: 'http://fleetserver:8220',
-            active: true,
-            is_default: true,
-            host_urls: ['http://fleetserver:8220'],
-          },
-        ],
-      } as any);
-      mockedListEnrollmentApiKeys.mockResolvedValue({
-        items: [
-          {
-            id: 'mocked-fleet-enrollment-token-id',
-            policy_id: 'mocked-fleet-enrollment-policy-id',
-            api_key: 'mocked-fleet-enrollment-api-key',
-          },
-        ],
-      } as any);
-
-      const createAgentlessAgentReturnValue = await agentPolicyService.createAgentlessAgent(
-        esClient,
-        soClient,
-        {
-          id: 'mocked-agentless-agent-policy-id',
-          name: 'agentless agent policy',
-          namespace: 'default',
-          supports_agentless: true,
-        } as AgentPolicy
-      );
-
-      expect(axios).toHaveBeenCalledTimes(1);
-      expect(createAgentlessAgentReturnValue).toEqual(returnValue);
-      expect(axios).toHaveBeenCalledWith(
-        expect.objectContaining({
-          data: expect.objectContaining({
-            fleet_token: 'mocked-fleet-enrollment-api-key',
-            fleet_url: 'http://fleetserver:8220',
-            policy_id: 'mocked-agentless-agent-policy-id',
-            stack_version: 'mocked-kibana-version-infinite',
-          }),
-          headers: expect.anything(),
-          httpsAgent: expect.anything(),
-          method: 'POST',
-          url: 'http://api.agentless.com/api/v1/ess/deployments',
-        })
-      );
-    });
   });
 
   describe('create', () => {
